requirements.txt
fastapi
python-telegram-bot==13.15
uvicorn
requests
telegram
python-dotenv